+++ src/tests/test_mime.py
@@ -1,7 +1,7 @@
 from __future__ import unicode_literals
 import pytest
 import six
-from autocrypt import header
+from autocrypt import mime
 from base64 import b64decode
 
 
@@ -14,31 +14,31 @@ def make_ac_dict(**kwargs):
     return d
 
 def test_parse_message_from_file(datadir):
-    msg = header.parse_message_from_file(datadir.open("rsa2048-simple.eml"))
+    msg = mime.parse_message_from_file(datadir.open("rsa2048-simple.eml"))
     assert msg.get_all("Autocrypt")
     assert msg.get_payload()
 
 def test_parse_message_from_string(datadir):
-    msg = header.parse_message_from_string(datadir.read("rsa2048-simple.eml"))
+    msg = mime.parse_message_from_string(datadir.read("rsa2048-simple.eml"))
     assert msg.get_all("Autocrypt")
     assert msg.get_payload()
 
 
 def test_make_and_parse_header_value():
     adr, keydata = "x@xy.z", "123"
-    h = header.make_ac_header_value(emailadr=adr, keydata=keydata)
-    d = header.parse_ac_headervalue(h)
-    assert not header.verify_ac_dict(d)
+    h = mime.make_ac_header_value(emailadr=adr, keydata=keydata)
+    d = mime.parse_ac_headervalue(h)
+    assert not mime.verify_ac_dict(d)
     assert d == make_ac_dict(to=adr, key=keydata)
 
 
 def test_make_and_parse_header_errors():
     adr, keydata = "x@xy.z", "123"
-    h = header.make_ac_header_value(
+    h = mime.make_ac_header_value(
         emailadr=adr, keydata=keydata, prefer_encrypt="notset", keytype="x")
     assert "prefer-encrypt" not in h, h
-    d = header.parse_ac_headervalue(h)
-    assert "unknown key type" in header.verify_ac_dict(d)[0]
+    d = mime.parse_ac_headervalue(h)
+    assert "unknown key type" in mime.verify_ac_dict(d)[0]
     assert d == make_ac_dict(to=adr, key=keydata, type="x")
 
 
@@ -69,12 +69,12 @@ class TestEmailCorpus:
 
     def test_rsa2048_unknown_critical(self, datadir):
         d = datadir.parse_ac_header_from_email("rsa2048-unknown-critical.eml")
-        l = header.verify_ac_dict(d)
+        l = mime.verify_ac_dict(d)
         assert len(l) == 1
         assert "unknown critical attr 'danger'" in l[0]
 
     def test_unknown_type(self, datadir):
         d = datadir.parse_ac_header_from_email("unknown-type.eml")
-        l = header.verify_ac_dict(d)
+        l = mime.verify_ac_dict(d)
         assert len(l) == 1
         assert "unknown key type 'x'" in l[0]
